backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
filelock==3.18.0
Flask==3.1.1
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
neo4j==5.28.1
networkx==3.4.2
numpy==2.2.5
pandas==2.2.3
propcache==0.3.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.7.1
six==1.17.0
SQLAlchemy==2.0.41
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
torch-geometric==2.6.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
